Removed php arrow tap dance
