+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/config.h
@@ -1,7 +1,11 @@
 #pragma once
 
 #define RGBLIGHT_LAYERS
-#define TAPPING_TOGGLE 3
+
+#define TAPPING_TOGGLE 2
 #define TAPPING_TERM 200
-#define AUTO_SHIFT_TIMEOUT 150
+
+#define AUTO_SHIFT_TIMEOUT 125
+#define NO_AUTO_SHIFT_ALPHA
+
 #define RGBLIGHT_LED_MAP {0, 1, 2, 3, 4, 5, 6, 7, 15, 14, 13, 12, 11, 10, 9, 8}
