+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/process_records.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "sharptsa.h"
+
+enum custom_keycodes {
+  RGB_DEFAULT, //Set RGB lighting to default
+};
