+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/rgb.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "sharptsa.h"
+
+void keyboard_post_init_rgb_light(void);
+void set_default_keyboard_rgb(void);
+void set_led_dance(void);
+void set_autoshift_disabled_rgb(void);
