+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/rules.mk
@@ -1,2 +1,7 @@
 TAP_DANCE_ENABLE = yes
 AUTO_SHIFT_ENABLE = yes
+
+SRC += sharptsa.c
+SRC += rgb.c
+SRC += tap_dance.c
+SRC += process_records.c
