+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/sharptsa.h
@@ -1,0 +1,6 @@
+#pragma once
+#include QMK_KEYBOARD_H
+
+#include "rgb.h"
+#include "tap_dance.h"
+#include "process_records.h"
