+++ keyboards/kbdfans/kbd75/keymaps/sharptsa/tap_dance.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "sharptsa.h"
+
+// Tap Dance declarations
+enum {
+    TD_RESET,
+    TD_LD
+};
+
+void td_led_dance_stop(void);
